serverfull-auctioneer/requirements.txt
fastapi
websockets
aiohttp
pyjwt
cryptography
fastapi_jwt_auth